Allez !

+++ SRC_checker/src/util.c
@@ -6,17 +6,17 @@
 /*   By: vloth <vloth@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/01 19:57:55 by vloth             #+#    #+#             */
-/*   Updated: 2022/03/07 02:54:58 by vloth            ###   ########.fr       */
+/*   Updated: 2022/03/18 23:32:19 by vloth            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../checker.h"
 
-int	ft_atoi(const char *nptr)
+int64_t	ft_atoi(const char *nptr)
 {
-	int	i;
-	int	n;
-	int	j;
+	int64_t	i;
+	int64_t	n;
+	int64_t	j;
 
 	j = 0;
 	i = 0;
